movies-info-service/src/test/java/unit/com/reactivespring/moviesinfoservice/controller/FluxAndMonoControllerTest.java
package com.reactivespring.moviesinfoservice.controller;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
import org.springframework.test.web.reactive.server.WebTestClient;
import reactor.core.publisher.Flux;
import reactor.test.StepVerifier;

import java.util.List;

import static org.junit.jupiter.api.Assertions.assertArrayEquals;
import static org.junit.jupiter.api.Assertions.assertEquals;

@WebFluxTest(controllers = FluxAndMonoController.class)
@AutoConfigureWebTestClient
class FluxAndMonoControllerTest {

    @Autowired
    WebTestClient webTestClient;

    @Test
    void fluxApproach1() {

        webTestClient
                .get()
                .uri("/flux")
                .exchange()
                .expectStatus()
                .is2xxSuccessful()
                .expectBodyList(Integer.class)
                .hasSize(5);

    }

    @Test
    void fluxApproach2() {

        Flux<Integer> res = webTestClient
                .get()
                .uri("/flux")
                .exchange()
                .expectStatus()
                .is2xxSuccessful()
                .returnResult(Integer.class)
                .getResponseBody();

        StepVerifier.create(res)
                .expectNext(1, 2, 3, 4, 5)
                .verifyComplete();

    }

    @Test
    void fluxApproach3() {

        webTestClient
                .get()
                .uri("/flux")
                .exchange()
                .expectStatus()
                .is2xxSuccessful()
                .expectBodyList(Integer.class)
                .consumeWith(listEntityExchangeResult -> {
                    List<Integer> res = listEntityExchangeResult.getResponseBody();
                    assert res != null;
                    assertArrayEquals(res.toArray(), List.of(1, 2, 3, 4, 5).toArray());
                });

    }

    @Test
    void mono() {

        webTestClient
                .get()
                .uri("/mono")
                .exchange()
                .expectStatus()
                .is2xxSuccessful()
                .expectBody(String.class)
                .consumeWith(stringEntityExchangeResult -> {
                    String actual = stringEntityExchangeResult.getResponseBody();
                    System.out.println(actual);
                    assertEquals("hello-world", actual);
                });

    }

    @Test
    void stream() {

        Flux<Long> res = webTestClient
                .get()
                .uri("/stream")
                .exchange()
                .expectStatus()
                .is2xxSuccessful()
                .returnResult(Long.class)
                .getResponseBody();

        StepVerifier.create(res)
                .expectNext(0L, 1L, 2L, 3L, 4L, 5L)
                .thenCancel()
                .verify();

    }
}